Fix issue where loss scaling was not applied in some cases.
When cloning=False was passed to Model.compile(), loss scaling was previously not applied.

In general, instances of LossScalingOptimizer do not scale the loss if tf.GradientTape/tf.gradients is used instead of calling optimizer.compute_gradients. In the future, we should probably find a way to warn or error out if the user creates a LossScaleOptimizer, then uses GradientTape without scaling the loss and gradients.

Also rename scale_grads to unscale_grads.

PiperOrigin-RevId: 247548487