Module-scoped buffer assignment.
- Use dataflow analysis in buffer assignment.

- Remove the notion of "Colocated Buffers" in buffer assignment to
  simplify the code, also remove the tests for them.

- Use dataflow analysis in memory space coloring (dependency of buffer assignment).

This cl brings us one step closer to the promised land without tuple
points to analysis and to the end of the state of split-brain
between dataflow analysis and tuple points to analysis.

It also greatly improves memory of RNN models with heavy use of
control flows and increases performance of some models by prioritizing
buffer reuse of DUS and scatters.

PiperOrigin-RevId: 248228644