Yet another performance improvement of LRNGrad on CPU builds
This improves the performance of the CPU version of LRNGrad by
skipping the innermost loop altogether when the gradient to be applied
at the end is *explicitly* zero (and thus nothing would change in the
output tensor). It turns out a lot of loops have 0 as value, and this
gives us a ~3% speed up on CIFAR-10 training on a Intel(R) Core(TM)
i9-7900X CPU @ 3.30GHz, with 20 threads (and remember that IEEE 754
defines the comparison 0.0==-0.0 to be true).