Use RefCountPtr in LookupResource to avoid leaks
LookupResource returns a raw pointer which the caller needs to Unref.
The prevalent pattern is this is followed by a ScopedUnref. This can be
problematic, since if a caller forgets to add a ScopedUnref call, we
have a memory leak. We resolve this by using RefCountPtr instead of a
raw pointer in LookupResource. Most use cases have been migrated in this
change.

Note some variables were renamed to handle line length restrictions.

PiperOrigin-RevId: 250423227