[XLA:CPU] Enable AVX512 when it's available
This was disabled in 2017 because LLVM's support was immature. It looks better
two years later so give it a try.

AVX512 is still an annoying beast because Skylake reduces the clock speed as
soon as you execute any AVX512 code. If the majority of your computation is not
dense AVX512 code this reduces performance quite a bit. However, MKL now always
targets AVX512 and Eigen does so if you compile it for Skylake, giving you a
slow clock rate without taking advantage of wider vectors in code generated
by XLA.

I ran a couple of benchmarks on Skylake with tensorflow and XLA built with
-march=skylake-avx512 and there are some nice performance wins.
PiperOrigin-RevId: 251193107