[XLA:GPU] Make TupleThunk faster by avoiding synchronous host/device callbacks.
TupleThunk registers a callback to free the host buffer (i.e. the CPU memory
that holds the pointers) after the H2D memcpy completes.  Unfortunately these
callbacks are slow.

We add a new routine to StreamExecutor which runs a callback on the next
BlockHostUntilDone, and we use this to free the buffer.  This means that the
buffer will stay live for a bit longer, but in practice these buffers are tiny,
less than 32 bytes.

We also use this in ConvolutionThunk and CudnnBatchnormThunk.  Previously these
were unsoundly relying on an optimization in the GPU driver (?) in which it
seemed to eagerly copy the host memory.

PiperOrigin-RevId: 251646672