Fix error when checkpointing Layer with "infer_float32_vars".
The error occurred when checkpointing an AutoCastVariable with a slot variable, if the variable name was different than the variable's field. E.g.:

def build(self, _):
  self.name2 = self.add_weight('name1', ())

In that case, the error message was "NotImplementedError: A slot variable was re-used as a dependency of a Trackable object."

The core issue was that when the variable name does not match the attribute name, the base Trackable class stores two dependencies to the same variable: one dependency with the variable name and one with the attribute name. This is normally not a problem, as the dependencies point to the same object and Checkpoint is smart enough to dedup the dependencies. However with AutoCastVariables before this change, one dependency would point to the base Variable, while the other would point to the AutoCastVariable, so both would be stored in the checkpoint. If the variable had a slot variable, both dependencies would have the same slot variable, confusing Checkpoint and causing it to throw the error.

This change causes both dependencies to point to the AutoCastVariable, fixing the issue.

PiperOrigin-RevId: 252454646