Refactor combinations.py to extract TF-specific parts out.
The logic is separated by introducing the TestCase class.  It may return instances of the Parameter class.

The code is also split into two files: one generic and one tensorflow-specific.  The eventual names I'd like to have are "combinations" and "tf_combinations" correspondingly, but I minimized the disruption of renaming to make this changelist easier to review.  I plan to do the rename separately.

It also now supports multiple arguments of type NamedDistribution.

PiperOrigin-RevId: 253119531